+++ Changes.md
@@ -20,6 +20,7 @@ Fixes
   - Fixed bugs which allowed read-only nodes to be edited.
   - Fixed undo for `Remove` menu item in Filter tab.
 - 3Delight : Fixed bug preventing 3delight from loading on Windows. (#5081). This also caused errors in the light editor at startup : `IECore.Exception: File "maya/osl/pointLight" could not be found.` [^2]
+- NodeEditor : Fixed bugs in handling of "green dot" non-default-value indicators with nested plugs.
 
 API
 ---